+++ src/mark.c
@@ -115,12 +115,12 @@ SEXP parse_gc_(SEXP x) {
   return out;
 }
 
-extern SEXP bench_max_memory_();
+extern SEXP bench_process_memory_();
 
 static const R_CallMethodDef CallEntries[] = {
     {"mark_", (DL_FUNC) &mark_, 5},
     {"system_time_", (DL_FUNC) &system_time_, 2},
-    {"bench_max_memory_", (DL_FUNC) &bench_max_memory_, 0},
+    {"bench_process_memory_", (DL_FUNC) &bench_process_memory_, 0},
     {"hires_time_", (DL_FUNC) &hires_time_, 0},
     {"parse_gc_", (DL_FUNC) &parse_gc_, 1},
     {NULL, NULL, 0}